authCtrl, authCtrl2.js 파일 병합, auth와 local.js의 기능 구분을 위한 일부 스웨거 이동 및 두 파일의 관리자기능 admin.js로 분리, 회원탈퇴 기능 추가(프론트 엔드 구현 안됨)
